requirements.txt
cryptography==2.6.1
flake8==3.7.8
SecretStorage==2.3.1
timeloop==1.0.2